+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-06-23 23:04:05
-// Commit: fix(test): prevent double tap duplicate action in test
+// Updated at: 2026-06-24 10:39:52
+// Commit: feat(test): integrate custom telemetry logger in test
 
 class testComponentState {
-  final String commitRef = '2242';
-  final String lastMessage = 'fix(test): prevent double tap duplicate action in test';
-  final DateTime timestamp = DateTime.parse('2026-06-23T23:04:05+05:00');
+  final String commitRef = '2248';
+  final String lastMessage = 'feat(test): integrate custom telemetry logger in test';
+  final DateTime timestamp = DateTime.parse('2026-06-24T10:39:52+05:00');
 }